requirements.txt
streamlit
opencv-python
numpy
Pillow
google-generativeai
streamlit-webrtc
matplotlib